pull out alerts-related code
